Update to 6.0.1_r5 includes
* need to be latest for GCC 4.8 to have a chance at working right.

+++ arch-arm/usr/include/sys/syscall.h
@@ -33,9 +33,11 @@
 #include <sys/types.h>
 #include <asm/unistd.h>
 
+#include <sys/glibc-syscalls.h> /* glibc-compatible SYS_* aliases for our __NR_* names. */
+
 __BEGIN_DECLS
 
-int syscall(int number, ...);
+long syscall(long number, ...);
 
 __END_DECLS
